+++ src/allocation-dispatcher.c
@@ -1,4 +1,4 @@
-#include <allocation_dispatcher.h>
+#include <allocation-dispatcher.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include <assert.h>
