+++ src/fixed-allocator.c
@@ -1,4 +1,4 @@
-#include <fixed_allocator.h>
+#include <fixed-allocator.h>
 #include <stdlib.h>
 #include <assert.h>
 #include <stdio.h>
